Fix tests with missing drives
