+++ Tests/Testably.Abstractions.Tests/FileSystemDriveInfoFactoryTests.cs
@@ -1,4 +1,5 @@
 using System.IO;
+using System.Linq;
 
 namespace Testably.Abstractions.Tests;
 
@@ -100,5 +101,64 @@ public abstract class FileSystemDriveInfoFactoryTests<TFileSystem>
         IFileSystem.IDriveInfo result = FileSystem.DriveInfo.Wrap(driveInfo);
 
         result.Name.Should().Be(driveInfo.Name);
+    }
+
+    [SkippableTheory]
+    [AutoData]
+    [FileSystemTests.DriveInfoFactory("MissingDrives")]
+    public void MissingDrive_WriteAllBytes_ShouldThrowDirectoryNotFoundException(
+        string path, byte[] contents)
+    {
+        Skip.IfNot(Test.RunsOnWindows);
+
+        IFileSystem.IDriveInfo driveInfo = GetUnmappedDrive();
+
+        path = $"{driveInfo.Name}{path}";
+
+        Exception? exception = Record.Exception(() =>
+        {
+            FileSystem.File.WriteAllBytes(path, contents);
+        });
+
+        exception.Should().BeOfType<DirectoryNotFoundException>()
+           .Which.Message.Should().Contain($"'{path}'");
+    }
+
+    [SkippableTheory]
+    [AutoData]
+    [FileSystemTests.DriveInfoFactory("MissingDrives")]
+    public void MissingDrive_CreateDirectoryInfo_ShouldOnlyThrowWhenAccessingData(
+        string path, string subPath)
+    {
+        Skip.IfNot(Test.RunsOnWindows);
+
+        IFileSystem.IDriveInfo driveInfo = GetUnmappedDrive();
+
+        path = $"{driveInfo.Name}{path}";
+        IFileSystem.IDirectoryInfo directoryInfo = FileSystem.DirectoryInfo.New(FileSystem.Path.Combine(path, subPath));
+        var parent = directoryInfo.Parent;
+
+        Exception? exception = Record.Exception(() =>
+        {
+            _ = parent!.EnumerateDirectories().ToArray();
+        });
+
+        exception.Should().BeOfType<DirectoryNotFoundException>()
+           .Which.Message.Should().Contain($"'{path}'");
+    }
+
+    private IFileSystem.IDriveInfo GetUnmappedDrive()
+    {
+        IFileSystem.IDriveInfo? driveInfo = null;
+        for (char c = 'A'; c <= 'Z'; c++)
+        {
+            driveInfo = FileSystem.DriveInfo.New($"{c}");
+            if (FileSystem.DriveInfo.GetDrives().All(d => d.Name != driveInfo.Name))
+            {
+                break;
+            }
+        }
+
+        return driveInfo ?? throw new NotImplementedException("No unmapped drive found!");
     }
 }
